Scripts to process data, CSV data
The first script, csv1, processes the Oct28 data, and was a test bed for
some ideas on how to massage the data. However, it's not complete for a
list of all of the members, so csv2 processes the TNL-Roster.. file,
which of course is missing the voice part information. And anyway, I'll
want to go back to the Oct28 file to get event sign-up information. So
this is just a 'stick it in the repo for now' type commit.